requirements.txt
amqp==2.6.1
asgiref==3.6.0
billiard==3.5.0.5
bleach==6.0.0
celery==4.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.0.0
dj-static==0.0.6
Django==3.2.19
django-anymail==10.0
django-bootstrap3==23.1
django-currentuser==0.5.3
django-debug-toolbar==4.1.0
django-import-export==3.2.0
django-localeurl==2.0.2
django-summernote==0.8.20.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
kombu==4.6.11
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pycparser==2.21
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vine==1.3.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
psycopg2>=2.6.1